adding location

+++ src/main/java/com/gege/activitypartner/dto/UserResponse.java
@@ -22,8 +22,8 @@ public class UserResponse {
     private Integer completedActivities;
     private String city; // User's city name
     private String placeId; // Google Places ID for the city
-    private BigDecimal latitude; // User's city latitude
-    private BigDecimal longitude; // User's city longitude
+    private Double latitude; // User's city latitude
+    private Double longitude; // User's city longitude
     private List<String> interests;
     private String badge;
     private List<UserPhotoResponse> photos = new ArrayList<>(); // User's photos (up to 6)
